remove old hle mouse
