+++ src/input/mouse.h
@@ -16,6 +16,14 @@ struct MouseState {
     s32 tilt = 0;
 };
 
+enum OrbisMouseButtonDataOffset {
+    ORBIS_MOUSE_BUTTON_PRIMARY = 0x00000001,
+    ORBIS_MOUSE_BUTTON_SECONDARY = 0x00000002,
+    ORBIS_MOUSE_BUTTON_OPTIONAL = 0x00000004,
+    ORBIS_MOUSE_BUTTON_OPTIONAL2 = 0x00000008,
+    ORBIS_MOUSE_BUTTON_OPTIONAL3 = 0x00000010,
+};
+
 constexpr u32 MAX_MOUSE_STATES = 64;
 
 class GameMouse {
